+++ src/com/madbros/tileminer/GameStates/CauldronStateInput.java
@@ -80,11 +80,11 @@ public class CauldronStateInput extends InventoryStateInput {
 	
 	public void additionalMouseMove() {
 		Rect mouseRect = new Rect(Helpers.getX(), Helpers.getY(), 1, 1);
-		Game.inventory.craftingMenu.handleMouseMove((int)mouseRect.x, (int)mouseRect.y);
+		Game.inventory.craftingMenu.handleMouseMove(mouseRect.x, mouseRect.y);
 		
 		CauldronTile cauldronTile = (CauldronTile) Game.level.activeBlocks[Game.inventory.currentInvActiveBlockX][Game.inventory.currentInvActiveBlockY].layers[OBJECT_LAYER];
 		if(cauldronTile.cauldronSlots[0].item.id != 0) {
-			cauldronTile.cauldronSlots[0].handleMouseMove((int)mouseRect.x, (int)mouseRect.y);
+			cauldronTile.cauldronSlots[0].handleMouseMove(mouseRect.x, mouseRect.y);
 		}
 		Slot[][] slots = {Game.inventory.invBar, Game.inventory.invBag, Game.inventory.craftingMenu.craftSlots, Game.inventory.invCrafted, Game.inventory.invClothing,cauldronTile.cauldronSlots, cauldronTile.fuelSlot, cauldronTile.craftedSlot};
 		for(int i = 0; i < slots.length; i++) {
