+++ src/com/madbros/tileminer/GameStates/ChestStateInput.java
@@ -70,7 +70,7 @@ public class ChestStateInput extends InventoryStateInput {
 	
 	public void additionalMouseMove() {
 		Rect mouseRect = new Rect(Helpers.getX(), Helpers.getY(), 1, 1);
-		Game.inventory.craftingMenu.handleMouseMove((int)mouseRect.x, (int)mouseRect.y);
+		Game.inventory.craftingMenu.handleMouseMove(mouseRect.x, mouseRect.y);
 		
 		//CauldronTile cauldronTile = (CauldronTile) Game.level.activeBlocks[Game.inventory.currentInvActiveBlockX][Game.inventory.currentInvActiveBlockY].layers[OBJECT_LAYER];
 		Slot[][] slots = {Game.inventory.invBar, Game.inventory.invBag, Game.inventory.craftingMenu.craftSlots, Game.inventory.invCrafted, Game.inventory.invClothing, Game.inventory.invChest};
