+++ src/com/madbros/tileminer/Items/Campfire.java
@@ -34,7 +34,7 @@ public class Campfire extends BlockItem32 {
 			if(Helpers.arrayDoesContainInt(placeableTileIds, block.getTopTerrainTile().id) && AIR == block.layers[OBJECT_LAYER].id) {
 
 				sprite.setColor(HIGHLIGHT_COLOR);
-				sprite.draw(x, y, Z_CHARACTER, (int)block.absRect.w, (int)block.absRect.h);
+				sprite.draw(x, y, Z_CHARACTER, block.absRect.w, block.absRect.h);
 				sprite.setColor(Color.WHITE);
 			}
 	}
