+++ src/com/madbros/tileminer/UI/TexturedUIButton.java
@@ -11,7 +11,7 @@ public class TexturedUIButton extends UIButton {
 	StaticSprite mainSprite = Sprites.buttonSprite;
 	StaticSprite pressedSprite = Sprites.pressedButtonSprite;
 	
-	public TexturedUIButton(double x, double y, double w, double h, String s, ButtonFunction func, SpriteBatch batch) {
+	public TexturedUIButton(int x, int y, int w, int h, String s, ButtonFunction func, SpriteBatch batch) {
 		super(x, y, w, h, s, func, batch);
 	}
